requirements.txt
openai
python-docx
pymupdf